implementing power editor

+++ src/store/modules/Login.js
@@ -34,7 +34,9 @@ const actions = {
 const mutations = {
     loginSuccess(state, response) {
         jsCookie.set('loggedUser', JSON.stringify({ ...response, token: response.type + ' ' + response.token, code: undefined, type: undefined }), { expires: 1 / 48 }); // 1/48 = 30min to expire
-        localStorage.setItem('role', response.roles[0]); // currently selected role
+        let roleStored = localStorage.getItem('role');
+        if (roleStored == null || !response.roles.includes(roleStored)) // if local stored did not exist
+            localStorage.setItem('role', response.roles[0]); // select first role in requested role list
         state.logined = true;
     },
     loginFailure(state) {
